[PONG:BASE] Refactor: Rectangle logic separated from paddle logic.
Paddle-related logic extracted to Paddle class. Paddle class is now concerned with all physics and gameplay mechanics, and extends Rectangle class, which is concerned with rectangle rendering.
