+++ pong/src/pong.ts
@@ -3,7 +3,7 @@ import { IDrawable } from "./interfaces.js";
 import { showError } from "./utils.js";
 import { AspectRatio } from "./types.js";
 import { Transform } from "./transform.js";
-import { Rectangle } from "./rectangle.js";
+import { Paddle } from "./paddle.js";
 import { Ball } from "./ball.js";
 import { Score } from "./score.js";
 import { VerticalDashedLine } from "./net.js";
@@ -45,8 +45,8 @@ class Pong
     private aspectRatio: AspectRatio;
     private backgroundColor: string;
     private net: VerticalDashedLine;
-    private leftPaddle: Rectangle;
-    private rightPaddle: Rectangle;
+    private leftPaddle: Paddle;
+    private rightPaddle: Paddle;
     private ball: Ball;
     private leftScore: Score;
     private rightScore: Score;
@@ -63,8 +63,8 @@ class Pong
         this.aspectRatio = aspectRatio;
         this.backgroundColor = backgroundColor;
         this.net = new VerticalDashedLine(new Transform( {x: Math.round(this.canvas.width * 0.5), y: Math.round(this.canvas.height * 0.5) }, 1), "black", 5, this.canvas.height, 10);
-        this.leftPaddle = new Rectangle(new Transform({ x: 100, y: Math.round(this.canvas.height * 0.5) }, 1), "black", 10, 100, 5, true);
-        this.rightPaddle = new Rectangle(new Transform({ x: this.canvas.width - 100, y: Math.round(this.canvas.height * 0.5) }, 1), "black", 10, 100, 5, true);
+        this.leftPaddle = new Paddle(new Transform({ x: 100, y: Math.round(this.canvas.height * 0.5) }, 1), "black", 10, 100, 5, true);
+        this.rightPaddle = new Paddle(new Transform({ x: this.canvas.width - 100, y: Math.round(this.canvas.height * 0.5) }, 1), "black", 10, 100, 5, true);
         this.ball = new Ball(new Transform({ x: Math.round(this.canvas.width * 0.5), y: Math.round(this.canvas.height * 0.5) }, 1), "white", 1, 10, true);
         this.leftScore = new Score(new Transform({ x: Math.round(this.canvas.width * 0.25), y: Math.round(this.canvas.height * 0.2) }, 1), "0", "white", 75);
         this.rightScore = new Score(new Transform({ x: Math.round(this.canvas.width * 0.75), y: Math.round(this.canvas.height * 0.2) }, 1), "0", "white", 75);
